Removed Hotspot interval 'Interval.dat' code
The hotspot sampling interval was originally read from the
'Interval.dat' file in mcpat.py. This file was created by
period-power.py.

Hotspot now uses the time period extracted by the partial argument.
This is stored in the variable `seconds`. This is more accurate since
it is also correct for the last sampling period which is often shorter
than the sampling period.