Added fix for unsigned overflow in timeJump var
If a task finishes and there are no tasks to schedule an offset is
calculated to move the arrival times of *all* tasks forward. The offset is
set so the earliest task can start running immediately.

There is an edge condition where SchedulerOpen::threadExit() is run
because a task finished and the earliest task to run has an arrival
time that smaller than the current time. So it should have already been
scheduled but the scheduler has not run yet. For example:

current time             336863
next_task_arrival_time   200000
scheduling epoch       10000000

This causes the unsigned timeJump to be come negative and thus overflow.
The adjustment results are still correct and unsigned values can
overflow so this is well-defined. But the log file will print a large
postive value instead of the correct negative one:

[Scheduler]: Readjusting Arrival Time by 18446744073709414753 ns

The fix it to make timeJump signed. It will then print:

[Scheduler]: Task arrival time in past, moving it to current time with negative arrival time adjustment.
[Scheduler]: Readjusting Arrival Time by -136863 ns